Move mdflow into about.dart and make private.
